abstraction of edit form
